Better idea: move them to util...
